requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Markdown==3.7
msgpack==1.1.0
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-engineio==4.10.1
python-socketio==5.11.4
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
simple-websocket==1.1.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==14.1
wsproto==1.2.0
zope.interface==7.2
